+++ paperless_vector_indexer/__init__.py
@@ -1,0 +1,13 @@
+from .client import Client
+from .models import SearchResult, Document
+from .exceptions import IndexerConnectionError, AuthenticationError, SearchError, DocumentNotFoundError
+
+__all__ = [
+    "Client",
+    "SearchResult",
+    "Document",
+    "IndexerConnectionError",
+    "AuthenticationError",
+    "SearchError",
+    "DocumentNotFoundError",
+]